requirements.txt
streamlit>=1.33.0
requests>=2.31.0
plotly>=5.18.0
websocket-client>=1.7.0